feat(p1): runtime plays a bundled $data map — LPC sprite rendering, collision, camera
The author→save→play loop closes: the Player now boots a bundled $data start map
(embedded), renders it with LPC tileset SPRITES, and you walk a character with tile
collision + a following camera — paint a map in the Studio (#16) and the runtime
renders + walks it.

  - D-0023: the repo IS the game — one repo = one game = one binary. No runtime
    project loader; the engine loads only its own bundled content (embedded,
    console-safe). A new game is a clone of the template.
  - Tileset moved Editor -> Engine.World (one source of truth for the editor AND
    the runtime renderer; TryGetTileIndex double->int to satisfy the MRM float ban).
  - New Engine.World.Camera (pure int centre+clamp) + Engine.Sim.StartMap
    (Build / BuildWorld / LoadWorld over MapSerializer + WorldSim, no events).
  - RpgGame: a fixed viewport, SpriteBatch tile blit via Tileset source-rects
    (replaces flat-color), a camera translation matrix, point-sampled crisp pixels.
  - A World/Entities/Data -/-> Engine.Sim NetArchTest rule guards the internal ring.

Gate GREEN [full]: 335 tests; coverage 93.0%; mutation Engine 86.33 (up) /
Abstractions 82.22 / Analyzers 91.18 / Editor 84.30.

Ticket #17 (403f50c5-aa6c-408c-b6fa-e540d17f425e).

+++ src/MonoRpgMaker.Studio/TilePalette.cs
@@ -6,6 +6,7 @@ using Avalonia.Input;
 using Avalonia.Media;
 using Avalonia.Media.Imaging;
 using MonoRpgMaker.Editor;
+using MonoRpgMaker.Engine.World;
 
 namespace MonoRpgMaker.Studio;
 
@@ -53,7 +54,7 @@ public sealed class TilePalette : Control
         base.OnPointerPressed(e);
 
         var pixel = e.GetPosition(this);
-        if (_session.Tileset.TryGetTileIndex(pixel.X, pixel.Y, out var index))
+        if (_session.Tileset.TryGetTileIndex((int)pixel.X, (int)pixel.Y, out var index))
         {
             _session.SelectTile(index, BlockingProvider?.Invoke() ?? false);
             InvalidateVisual();
